[IMCC] Fixed a few more signedness warnings for set functions as fallout from r28364.
git-svn-id: https://svn.parrot.org/parrot/trunk@28367 d31e2699-5ff4-0310-a27c-f18f2fbe73fe

+++ compilers/imcc/cfg.c
@@ -1193,8 +1193,8 @@ sort_loops(PARROT_INTERP, ARGIN(IMC_Unit *unit))
     int          changed;
     Loop_info   *li;
 
-    int         n_loops   = unit->n_loops;
-    Loop_info **loop_info = unit->loop_info;
+    unsigned int n_loops  = (unsigned int)unit->n_loops;
+    Loop_info  **loop_info = unit->loop_info;
 
     for (i = 0; i < n_loops; i++) {
         loop_info[i]->size = 0;
